updated code based on original instructions
