+++ .gitignore
@@ -128,3 +128,6 @@ iframe_figures/figure_204.html
 iframe_figures/figure_207.html
 iframe_figures/figure_212.html
 iframe_figures/figure_213.html
+explore_data.ipynb
+.DS_Store
+.vscode/settings.json
